Update mod_congratsonlevel.cpp
Fixed a logical error in the language prompts. The prompt "1 gold coin = 10000 copper coins" incorrectly displayed "You've been awarded " << money << " gold and a few treasures!" It should be corrected to "You've been awarded " << money << " Copper and a few treasures!" The actual currency unit issued in the on-screen prompt is Copper. The Spanish prompt should be corrected to: cobre.